fix: remove old conflict detection mechanism
