+++ packages/salesforcedx-vscode-core/test/vscode-integration/commands/forceSourceDiff.test.ts
@@ -14,7 +14,7 @@ import { commands, Uri } from 'vscode';
 import { channelService } from '../../../src/channels';
 import { forceSourceDiff } from '../../../src/commands';
 import * as conflictCommands from '../../../src/commands';
-import * as conflictDetectionService from '../../../src/conflict/conflictDetectionService';
+import * as differ from '../../../src/conflict/directoryDiffer';
 import {
   MetadataCacheResult,
   MetadataCacheService,
@@ -194,8 +194,8 @@ describe('Force Source Diff', () => {
 
     beforeEach(() => {
       notificationStub = stub(notificationService, 'showErrorMessage');
-      diffOneFileStub = stub(conflictDetectionService, 'diffOneFile');
-      diffFolderStub = stub(conflictDetectionService, 'diffFolder');
+      diffOneFileStub = stub(differ, 'diffOneFile');
+      diffFolderStub = stub(differ, 'diffFolder');
     });
 
     afterEach(() => {
