+++ packages/salesforcedx-vscode-core/test/vscode-integration/commands/util/postconditionCheckers.test.ts
@@ -14,7 +14,6 @@ import { join, normalize } from 'path';
 import { createSandbox, SinonSandbox, SinonStub } from 'sinon';
 import { channelService } from '../../../../src/channels';
 import {
-  ConflictDetectionChecker,
   ConflictDetectionMessages,
   EmptyPostChecker,
   OverwriteComponentPrompt,
@@ -22,7 +21,6 @@ import {
 } from '../../../../src/commands/util';
 import { TimestampConflictChecker } from '../../../../src/commands/util/postconditionCheckers';
 import {
-  conflictDetector,
   conflictView,
   DirectoryDiffResults
 } from '../../../../src/conflict';
@@ -292,170 +290,9 @@ describe('Postcondition Checkers', () => {
     }
   });
 
-  describe('ConflictDetectionChecker', () => {
-    let modalStub: SinonStub;
-    let settingsStub: SinonStub;
-    let detectorStub: SinonStub;
-    let conflictViewStub: SinonStub;
-    let appendLineStub: SinonStub;
-    let channelOutput: string[] = [];
-
-    beforeEach(() => {
-      env = createSandbox();
-      channelOutput = [];
-      modalStub = env.stub(notificationService, 'showWarningModal');
-      settingsStub = env.stub(sfdxCoreSettings, 'getConflictDetectionEnabled');
-      detectorStub = env.stub(conflictDetector, 'checkForConflicts');
-      conflictViewStub = env.stub(conflictView, 'visualizeDifferences');
-      appendLineStub = env.stub(channelService, 'appendLine');
-      appendLineStub.callsFake(line => channelOutput.push(line));
-    });
-
-    afterEach(() => env.restore());
-
-    const emptyMessages: ConflictDetectionMessages = {
-      warningMessageKey: '',
-      commandHint: i => i
-    };
-
-    const retrieveMessages: ConflictDetectionMessages = {
-      warningMessageKey: 'conflict_detect_conflicts_during_retrieve',
-      commandHint: i => i
-    };
-
-    const validInput: ContinueResponse<string> = {
-      type: 'CONTINUE',
-      data: 'package.xml'
-    };
-
-    it('Should return CancelResponse if input passed in is CancelResponse', async () => {
-      const postChecker = new ConflictDetectionChecker(emptyMessages);
-      const response = await postChecker.check({ type: 'CANCEL' });
-      expect(response.type).to.equal('CANCEL');
-    });
-
-    it('Should return ContinueResponse unchanged if input is ContinueResponse & conflict detection is disabled', async () => {
-      const postChecker = new ConflictDetectionChecker(emptyMessages);
-
-      settingsStub.returns(false);
-      const response = await postChecker.check(validInput);
-
-      expect(response.type).to.equal('CONTINUE');
-      if (response.type === 'CONTINUE') {
-        expect(response.data).to.equal('package.xml');
-      } else {
-        expect.fail('Response should be of type ContinueResponse');
-      }
-    });
-
-    it('Should return CancelResponse when a username is not defined.', async () => {
-      const postChecker = new ConflictDetectionChecker(emptyMessages);
-      settingsStub.returns(true);
-
-      const response = await postChecker.check(validInput);
-      expect(response.type).to.equal('CANCEL');
-    });
-
-    it('Should return ContinueResponse when no conflicts are detected', async () => {
-      const postChecker = new ConflictDetectionChecker(emptyMessages);
-      const response = await postChecker.handleConflicts(
-        'manifest.xml',
-        'admin@example.com',
-        { different: new Set<string>() } as DirectoryDiffResults
-      );
-
-      expect(response.type).to.equal('CONTINUE');
-      expect((response as ContinueResponse<string>).data).to.equal(
-        'manifest.xml'
-      );
-      expect(appendLineStub.notCalled).to.equal(true);
-    });
-
-    it('Should post a warning and return CancelResponse when conflicts are detected and cancelled', async () => {
-      const postChecker = new ConflictDetectionChecker(retrieveMessages);
-      const results = {
-        different: new Set<string>([
-          'main/default/objects/Property__c/fields/Broker__c.field-meta.xml',
-          'main/default/aura/auraPropertySummary/auraPropertySummaryController.js'
-        ]),
-        scannedLocal: 4,
-        scannedRemote: 6
-      } as DirectoryDiffResults;
-      modalStub.returns('Cancel');
-
-      const response = await postChecker.handleConflicts(
-        'package.xml',
-        'admin@example.com',
-        results
-      );
-      expect(response.type).to.equal('CANCEL');
-
-      expect(modalStub.firstCall.args.slice(1)).to.eql([
-        nls.localize('conflict_detect_override'),
-        nls.localize('conflict_detect_show_conflicts')
-      ]);
-
-      expect(channelOutput).to.include.members([
-        nls.localize('conflict_detect_conflict_header', 2, 6, 4),
-        normalize(
-          'main/default/objects/Property__c/fields/Broker__c.field-meta.xml'
-        ),
-        normalize(
-          'main/default/aura/auraPropertySummary/auraPropertySummaryController.js'
-        ),
-        nls.localize('conflict_detect_command_hint', 'package.xml')
-      ]);
-
-      expect(conflictViewStub.calledOnce).to.equal(true);
-    });
-
-    it('Should post a warning and return ContinueResponse when conflicts are detected and overwritten', async () => {
-      const postChecker = new ConflictDetectionChecker(retrieveMessages);
-      const results = {
-        different: new Set<string>('MyClass.cls')
-      } as DirectoryDiffResults;
-      modalStub.returns(nls.localize('conflict_detect_override'));
-
-      const response = await postChecker.handleConflicts(
-        'manifest.xml',
-        'admin@example.com',
-        results
-      );
-      expect(response.type).to.equal('CONTINUE');
-
-      expect(modalStub.firstCall.args.slice(1)).to.eql([
-        nls.localize('conflict_detect_override'),
-        nls.localize('conflict_detect_show_conflicts')
-      ]);
-    });
-
-    it('Should post a warning and return CancelResponse when conflicts are detected and conflicts are shown', async () => {
-      const postChecker = new ConflictDetectionChecker(retrieveMessages);
-      const results = {
-        different: new Set<string>('MyClass.cls')
-      } as DirectoryDiffResults;
-      modalStub.returns(nls.localize('conflict_detect_show_conflicts'));
-
-      const response = await postChecker.handleConflicts(
-        'manifest.xml',
-        'admin@example.com',
-        results
-      );
-      expect(response.type).to.equal('CANCEL');
-
-      expect(modalStub.firstCall.args.slice(1)).to.eql([
-        nls.localize('conflict_detect_override'),
-        nls.localize('conflict_detect_show_conflicts')
-      ]);
-
-      expect(conflictViewStub.calledOnce).to.equal(true);
-    });
-  });
-
   describe('TimestampConflictChecker', () => {
       let modalStub: SinonStub;
       let settingsStub: SinonStub;
-      let detectorStub: SinonStub;
       let conflictViewStub: SinonStub;
       let appendLineStub: SinonStub;
       let channelOutput: string[] = [];
@@ -465,7 +302,6 @@ describe('Postcondition Checkers', () => {
         channelOutput = [];
         modalStub = env.stub(notificationService, 'showWarningModal');
         settingsStub = env.stub(sfdxCoreSettings, 'getConflictDetectionEnabled');
-        detectorStub = env.stub(conflictDetector, 'checkForConflicts');
         conflictViewStub = env.stub(conflictView, 'visualizeDifferences');
         appendLineStub = env.stub(channelService, 'appendLine');
         appendLineStub.callsFake(line => channelOutput.push(line));
